move translate & combine fn to hir
