+++ src/utils.rs
@@ -62,30 +62,7 @@ pub(crate) fn split_trees(text: &str) -> Vec<&str> {
 /// This function is called during verification of the combined HIR.
 pub(crate) fn get_contract_name_from_identifier(identifier: &String) -> String {
     let identifier_parts: Vec<&str> = identifier.split(CONTRACT_PART_SEPARATOR).collect();
-    return identifier_parts[0].to_string();
-}
-
-/// Generates the HIR for a single tree.
-///
-/// This function leverages `crate::syntax::parse` and `crate::hir::translator::Translator::translate`
-/// to hide away most of the complexity of `bulloak`'s internal compiler.
-pub(crate) fn translate_tree_to_hir(tree: &str) -> crate::error::Result<crate::hir::Hir> {
-    let ast = crate::syntax::parse(tree)?;
-    let mut discoverer = crate::scaffold::modifiers::ModifierDiscoverer::new();
-    let modifiers = discoverer.discover(&ast);
-    Ok(crate::hir::translator::Translator::new().translate(&ast, modifiers))
-}
-
-/// High-level function that returns a HIR given the contents of a `.tree` file.
-///
-/// This function leverages `translate_tree_to_hir` to generate the HIR for each tree,
-/// and `crate::hir::combiner::Combiner::combine` to combine the HIRs into a single HIR.
-pub(crate) fn translate_and_combine_trees(text: &str) -> crate::error::Result<crate::hir::Hir> {
-    let trees = split_trees(text);
-    let hirs = trees.iter()
-        .map(|tree| translate_tree_to_hir(tree))
-        .collect::<crate::error::Result<Vec<crate::hir::Hir>>>()?;
-    Ok(crate::hir::combiner::Combiner::new().combine(&hirs)?)
+    return sanitize(identifier_parts[0]);
 }
